One more case

+++ BST.py
@@ -114,6 +114,10 @@ def delete(root,val):
     root = delete(root,x)
     """
     
+    #Searching twice but for now its fine
+    if not search(root,val):
+        raise ValueNotPresentException
+    
     if root.val == val:
         return delete_helper(root)
     elif root.val < val:
